pytest_demo/requirements.txt
pytest==7.2.0
requests==2.27.1
PyHamcrest==2.0.4
selenium==4.12.0